+++ src/IV.DX/IV.DX.Hosting/DXExecutionContextResolver.cs
@@ -1,3 +1,4 @@
+using IV.DX.Kernel.Enums;
 using IV.DX.Kernel.Models;
 using IV.DX.Persistence.Contracts.Abstractions;