+++ src/IV.DX/IV.DX.Persistence/DXRawReader.cs
@@ -1,4 +1,5 @@
 using IV.DX.Kernel;
+using IV.DX.Kernel.Enums;
 using IV.DX.Kernel.Helpers;
 using IV.DX.Kernel.Models;
 using IV.DX.Persistence.Contracts.Abstractions;