+++ src/IV.DX/Tests/IntTests/IV.DX.Application.IntTests/Services/DXSecurityInitializationModeTests.cs
@@ -1,3 +1,4 @@
+using IV.DX.Kernel.Enums;
 using IV.DX.PostgreSQL;
 using IV.DX.Application.Contracts.Abstractions;
 using IV.DX.Hosting;